mmv/mmv/requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
ffmpeg-python==0.2.0
future==0.18.2
llvmlite==0.33.0
numba==0.50.1
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pycparser==2.20
PyYAML==5.3.1
resampy==0.2.2
scipy==1.5.1
six==1.15.0
SoundFile==0.10.3.post1
svgwrite==1.4
tinycss2==1.0.2
Wand==0.6.2
webencodings==0.5.1
